[CP-2425] App structure refactored
