+++ libs/core/core/components/apps/base-app/base-app.component.tsx
