+++ libs/core/core/components/apps/privacy-policy-app.component.tsx
@@ -4,21 +4,20 @@
  */
 
 import React from "react"
-import history from "Core/core/routes/history"
+import history from "Core/core/history"
 import { Route, Router } from "react-router"
 import { FunctionComponent } from "Core/core/types/function-component.interface"
 import { URL_MAIN } from "Core/__deprecated__/renderer/constants/urls"
-import { TermsOfServiceUI } from "Core/settings/components"
+import { PrivacyPolicyUI } from "Core/settings/components"
 
-
-const TermsOfServiceApp: FunctionComponent = ( ) => {
+const PrivacyPolicyApp: FunctionComponent = ( ) => {
   return (
     <Router history={history}>
-      <Route path={URL_MAIN.termsOfService}>
-        <TermsOfServiceUI />
+      <Route path={URL_MAIN.privacyPolicy}>
+        <PrivacyPolicyUI />
       </Route>
     </Router>
   )
 }
 
-export default TermsOfServiceApp
+export default PrivacyPolicyApp
