+++ libs/core/core/components/root-wrapper.tsx
@@ -3,68 +3,20 @@
  * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
  */
 
-import React, { useMemo } from "react"
+import React from "react"
 import { IntlProvider } from "react-intl"
-import { History } from "history"
-import { ipcRenderer } from "electron-better-ipc"
 import translationConfig from "App/translations.config.json"
 import { ThemeProvider } from "styled-components"
 import { Normalize } from "styled-normalize"
 import GlobalStyle from "Core/core/styles/global-style.component"
 import theme from "Core/core/styles/theming/theme"
 import { FunctionComponent } from "Core/core/types/function-component.interface"
-import HelpApp from "Core/core/components/help-app.component"
-import ErrorApp from "Core/core/components/error-app.component"
-import BaseAppContainer from "Core/core/components/base-app-container.component"
-import { QuestionAndAnswer } from "Core/help/components/help.component"
-import LicenseApp from "Core/core/components/license-app.component"
-import TermsOfServiceApp from "Core/core/components/terms-of-service-app.component"
-import PrivacyPolicyApp from "Core/core/components/privacy-policy-app.component"
-import SarApp from "Core/core/components/sar-app.component"
-import { HelpActions } from "Core/__deprecated__/common/enums/help-actions.enum"
-import { Mode } from "Core/__deprecated__/common/enums/mode.enum"
 import localeEn from "Core/__deprecated__/renderer/locales/default/en-US.json"
 import { ModalProvider } from "Core/__deprecated__/renderer/components/core/modal/modal.service"
 import modalService from "Core/__deprecated__/renderer/components/core/modal/modal.service"
+import AppsSwitch from "Core/core/components/apps-switch"
 
 const RootWrapper: FunctionComponent = () => {
-  const mode = new URLSearchParams(window.location.search).get("mode")
-  const saveToStore = async (normalizeData: QuestionAndAnswer) =>
-    await ipcRenderer.callMain(HelpActions.SetStoreValue, normalizeData)
-  const getStoreData = async (key?: string) =>
-    await ipcRenderer.callMain(HelpActions.GetStore, key)
-
-  const RenderRoutes = useMemo(
-    () => () => {
-      if (mode === Mode.ServerError) {
-        return <ErrorApp />
-      }
-
-      if (mode === Mode.Help) {
-        return <HelpApp saveToStore={saveToStore} getStoreData={getStoreData} />
-      }
-
-      if (mode === Mode.License) {
-        return <LicenseApp />
-      }
-
-      if (mode === Mode.TermsOfService) {
-        return <TermsOfServiceApp />
-      }
-
-      if (mode === Mode.PrivacyPolicy) {
-        return <PrivacyPolicyApp />
-      }
-
-      if (mode === Mode.Sar) {
-        return <SarApp />
-      }
-
-      return <BaseAppContainer />
-    },
-    [mode]
-  )
-
   return (
     <ThemeProvider theme={theme}>
       <IntlProvider
@@ -75,7 +27,7 @@ const RootWrapper: FunctionComponent = () => {
         <ModalProvider service={modalService}>
           <Normalize />
           <GlobalStyle />
-          <RenderRoutes />
+          <AppsSwitch />
         </ModalProvider>
       </IntlProvider>
     </ThemeProvider>
