+++ libs/core/update/components/update-os-interrupted-flow/update-os-interrupted-flow.test.tsx
@@ -10,12 +10,14 @@ import { UpdateOsInterruptedFlowProps } from "Core/update/components/update-os-i
 import { renderWithThemeAndIntl } from "Core/__deprecated__/renderer/utils/render-with-theme-and-intl"
 import { URL_ONBOARDING } from "Core/__deprecated__/renderer/constants/urls"
 
+// AUTO DISABLED - fix me if you like :)
+// eslint-disable-next-line @typescript-eslint/no-unsafe-return
 jest.mock('react-router-dom', () => ({
   ...jest.requireActual('react-router-dom'),
-  useHistory: (): History => ({
+  useHistory: () => ({
     push: jest.fn(),
     location: { pathname: URL_ONBOARDING.welcome },
-  } as unknown as History),
+  }),
 }));
 
 const defaultProps: UpdateOsInterruptedFlowProps = {
